+++ graph/graph.ts
@@ -2,63 +2,70 @@ namespace $ {
 	
 	export class $mol_graph< Node , Edge > {
 		
-		nodes : { [ id : string ] : Node | undefined } = {}
+		nodes = new Set< Node >()
 		
-		edgesOut : { [ from : string ] : { [ to : string ] : Edge } } = {}
-		edgesIn : { [ to : string ] : { [ from : string ] : Edge } } = {}
+		edges_out = new Map< Node , Map< Node , Edge > >()
+		edges_in = new Map< Node , Map< Node , Edge > >()
 		
-		nodeEnsure( id : string ) {
-			if( this.nodes.hasOwnProperty( id ) ) return
-			this.nodes[ id ] = undefined
-		}
-		
-		linkOut( from : string , to : string , edge : Edge ) {
-			if( !this.edgesOut[ from ] ) {
-				this.edgesOut[ from ] = {}
-				this.nodeEnsure( from )
-			}
-			this.edgesOut[ from ][ to ] = edge
-			this.nodeEnsure( to )
-		}
-		
-		linkIn( to : string , from : string , edge : Edge ) {
-			if( !this.edgesIn[ to ] ) {
-				this.edgesIn[ to ] = {}
-				this.nodeEnsure( to )
-			}
-			this.edgesIn[ to ][ from ] = edge
-			this.nodeEnsure( from )
-		}
-		
-		edgeOut( from : string , to : string ) {
-			return this.edgesOut[ from ] && this.edgesOut[ from ][ to ]
-		}
-		
-		edgeIn( to : string , from : string ) {
-			return this.edgesIn[ to ] && this.edgesIn[ to ][ from ]
-		}
-		
-		link( from : string , to : string , edge : Edge ) {
-			this.linkOut( from , to , edge )
-			this.linkIn( to , from , edge )
-		}
-		
-		unlink( from : string , to : string ) {
-			delete this.edgesIn[ to ][ from ]
-			delete this.edgesOut[ from ][ to ]
-		}
-		
-		cut_cycles( get_weight : ( edge : Edge )=> number ) {
+		link_out( from : Node , to : Node , edge : Edge ) {
 			
-			const checked = [] as string[]
+			let pair = this.edges_out.get( from )
 			
-			for( const start in this.nodes ) {
-				
-				const path = [] as string[]
-				
-				const visit = ( from : string ) : number => {
+			if( !pair ) {
+				pair = new Map< Node , Edge >()
+				this.edges_out.set( from , pair )
+				this.nodes.add( from )
+			}
+			
+			pair.set( to , edge )
+			this.nodes.add( to )
 
-					if( checked.includes( from ) ) return Number.POSITIVE_INFINITY
+		}
+		
+		link_in( to : Node , from : Node , edge : Edge ) {
+
+			let pair = this.edges_in.get( from )
+			
+			if( !pair ) {
+				pair = new Map< Node , Edge >()
+				this.edges_in.set( from , pair )
+				this.nodes.add( from )
+			}
+			
+			pair.set( to , edge )
+			this.nodes.add( to )
+
+		}
+		
+		edge_out( from : Node , to : Node ) {
+			return this.edges_out.get( from )?.get( to ) ?? null
+		}
+		
+		edge_in( to : Node , from : Node ) {
+			return this.edges_in.get( to )?.get( from ) ?? null
+		}
+		
+		link( from : Node , to : Node , edge : Edge ) {
+			this.link_out( from , to , edge )
+			this.link_in( to , from , edge )
+		}
+		
+		unlink( from : Node , to : Node ) {
+			this.edges_in.get( to )?.delete( from )
+			this.edges_out.get( from )?.delete( to )
+		}
+		
+		acyclic( get_weight : ( edge : Edge )=> number ) {
+			
+			const checked = [] as Node[]
+			
+			for( const start of this.nodes ) {
+				
+				const path = [] as Node[]
+				
+				const visit = ( from : Node ) : number => {
+
+					if( checked.includes( from ) ) return Number.MAX_SAFE_INTEGER
 
 					const index = path.lastIndexOf( from )
 					if( index > -1 ) {
@@ -66,28 +73,30 @@ namespace $ {
 						const cycle = path.slice( index )
 
 						return cycle.reduce(
-							( weight , id , index )=> Math.min(
+							( weight , node , index )=> Math.min(
 								weight ,
-								get_weight( this.edgesOut[ id ][ cycle[ ( index + 1 ) % cycle.length ] ] ) ,
+								get_weight( this.edge_out( node , cycle[ ( index + 1 ) % cycle.length ] )! ) ,
 							) ,
-							Number.POSITIVE_INFINITY ,
+							Number.MAX_SAFE_INTEGER ,
 						)
 
 					}
 
 					path.push( from )
 
-					try {
+					dive: try {
 
-						const deps = this.edgesOut[ from ]
-						for( const to in deps ) {
+						const deps = this.edges_out.get( from )
+						if( !deps ) break dive
+
+						for( const [ to , edge ] of deps ) {
 
 							if( to === from ) {
 								this.unlink( from , to )
 								continue
 							}
 
-							const weight_out = get_weight( deps[ to ] )
+							const weight_out = get_weight( edge )
 							const min = visit( to )
 							
 							if( weight_out > min ) return min
@@ -101,7 +110,7 @@ namespace $ {
 
 					checked.push( from )
 
-					return Number.POSITIVE_INFINITY
+					return Number.MAX_SAFE_INTEGER
 				}
 
 				visit( start )
@@ -112,21 +121,24 @@ namespace $ {
 		
 		get sorted() {
 
-			const sorted : string[] = []
+			const sorted = new Set< Node >()
 			
-			const visit = ( id : string ) => {
+			const visit = ( node : Node ) => {
 				
-				if( sorted.indexOf( id ) !== -1 ) return
+				if( sorted.has( node ) ) return
 
-				for( const dep in this.edgesOut[ id ] ) visit( dep )
+				const deps = this.edges_out.get( node )
 
-				if( sorted.indexOf( id ) !== -1 ) return
+				if( deps ) {
+					for( const [dep] of deps ) visit( dep )
+				}
 
-				sorted.push( id )
-
+				sorted.add( node )
 			}
 			
-			Object.keys( this.nodes ).forEach( id => visit( id ) )
+			for( const node of this.nodes ) {
+				visit( node )
+			}
 			
 			return sorted
 		}
